requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
billiard==4.2.4
celery==5.6.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.128.0
flake8==7.3.0
frozenlist==1.8.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.6.2
langchain==1.2.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.5
langchain-openai==1.1.6
langchain-qdrant==1.1.0
langchain-text-splitters==1.1.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.5.2
marshmallow==3.26.2
mccabe==0.7.0
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.4.0
openai==2.14.0
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pluggy==1.6.0
portalocker==3.2.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
PyYAML==6.0.3
qdrant-client==1.16.2
redis==7.1.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.45
starlette==0.50.0
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
uuid_utils==0.12.0
uvicorn==0.40.0
vine==5.1.0
wcwidth==0.2.14
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
